Link encapsulated into separate component and implemented for 'Show more' and 'Set price'
