+++ src/v2/Components/ArtworkFilter/ArtworkFilters/ShowMore.tsx
@@ -1,5 +1,5 @@
-import { Clickable, Text, TextVariant, useThemeConfig } from "@artsy/palette"
 import React, { Children, isValidElement, useState } from "react"
+import { FilterLink } from "./FilterLink"
 
 interface ShowMoreProps {
   initial?: number
@@ -19,26 +19,14 @@ export const ShowMore: React.FC<ShowMoreProps> = ({
 
   const [isExpanded, setExpanded] = useState(expanded)
 
-  const tokens = useThemeConfig({
-    v2: { variant: "small" as TextVariant },
-    v3: { variant: "xs" as TextVariant },
-  })
-
   return (
     <>
       {isExpanded ? nodes : nodes.slice(0, initial)}
 
       {hasMore && (
-        <Clickable
-          mt={1}
-          textDecoration="underline"
-          textAlign="left"
-          onClick={() => setExpanded(visibility => !visibility)}
-        >
-          <Text variant={tokens.variant}>
-            {isExpanded ? "Hide" : "Show more"}
-          </Text>
-        </Clickable>
+        <FilterLink onClick={() => setExpanded(visibility => !visibility)}>
+          {isExpanded ? "Hide" : "Show more"}
+        </FilterLink>
       )}
     </>
   )
